+++ app.js
@@ -5,6 +5,7 @@ var cookieParser = require('cookie-parser');
 var logger = require('morgan');
 var expressHbs = require('express-handlebars');
 var indexRouter = require('./routes/index');
+var userRoutes = require('./routes/user');
 var mongoose = require('mongoose');
 var app = express();
 var session = require('express-session');
@@ -22,17 +23,21 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended:false }));
 app.use(validator());
 app.use(flash());
+app.use(session({secret: 'wd6secret', resave: false, saveUnitialized: false}));
 app.use(passport.initialize());
 app.use(passport.session());
 app.use(logger('dev'));
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(session({secret: 'wd6secret', resave: false, saveUnitialized: false}));
 app.use(express.static(path.join(__dirname, 'public')));
 
+app.use(function (req,res,next) {
+  res.locals.login = req.isAuthenticated();
+  next();
+});
+app.use('/user', userRoutes);
 app.use('/', indexRouter);
-
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
