+++ views/partials/header.hbs
@@ -1,10 +1,10 @@
 <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-    <a class="navbar-brand" href="#">WD6</a>
+    <a class="navbar-brand" href="/">WD6</a>
     <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
         <span class="navbar-toggler-icon"></span>
     </button>
 
-    <div class="collapse navbar-collapse" id="navbarSupportedContent">
+    <ul class="collapse navbar-collapse" id="navbarSupportedContent">
         <ul class="navbar-nav mr-auto">
             <li class="nav-item active">
                 <a class="nav-link" href="#">Home <span class="sr-only">(current)</span></a>
@@ -17,8 +17,21 @@
         <li class="nav-item">
             <a class="nav-link" href="#"><i class="fas fa-shopping-cart"></i> Cart</a>
         </li>
+            <li class="nav-item dropdown">
+                <a class="nav-link dropdown-toggle" data-toggle="dropdown" href="#" role="button" aria-haspopup="true" aria-expanded="false" ><i class="fas fa-user"></i> User Management</a>
+                <ul class="dropdown-menu">
+                    {{# if login}}
+                        <li class="dropdown-item"><a href="/user/profile">User Account</a></li>
+                        <li role="separator" class="divider"></li>
+                        <li class="dropdown-item"><a href="/user/logout">Logout</a></li>
+                        {{ else }}
+                        <li class="dropdown-item"><a href="/user/signup">Sign Up</a></li>
+                        <li class="dropdown-item"><a href="/user/signin">Sign In</a></li>
+                    {{/if}}
+                </ul>
+            </li>
         <li class="nav-item">
-            <a class="nav-link" href="#"><i class="fas fa-user"></i> User Management</a>
+
         </li>
         </ul>
     </div>
