+++ views/user/signin.hbs
@@ -19,7 +19,7 @@
                     <input type="password" id="password" name="password" class="form-control">
                 </div>
                 <input type="hidden" name="_csrf" value="{{ csrfToken }}">
-                <button type="submit" class="btn btn-primary">Sign Up</button>
+                <button type="submit" class="btn btn-primary">Sign In</button>
             </form>
         </div>
     </div>
